pyproject.toml
[project]
name = "kanban"
version = "0.1.0"
description = "치과 상담 예약 칸반 풀스택 앱"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",
    "psycopg2-binary>=2.9",
    "python-dotenv>=1.0",
    "pydantic>=2.9",
]

[tool.uv]
package = false
